+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to QETLAB will be documented in this file.
 - ParallelRepetition: Produces the coefficients of a parallel repetition of a nonlocal game, in full probability notation. Replaces the old NonlocalGameValue.m.
 - PolynomialAsMatrix: Creates a compact fully symmetric matrix representation of a polynomial.
 - PolynomialOptimize: Bounds the optimal value of a homogeneous polynomial on the unit sphere.
+- PolynomialSOS: Bounds the optimal value of a homogeneous polynomial on the unit sphere via the Sum-Of-Squares hierarchy.
 - RandomPPTState: Generates a random density matrix with positive partial transpose, and optionally low rank.
 - helpers/asum_vector: Creates all vectors with binary entries adding to a given value. Used by AntisymmetricProjection.m.
 - helpers/asymind: Creates all vectors with strictly increasing permutations of an input vector.